Update SMTP Library and Settings
Changed const char* to String in the SMTP library and enabled it in the
main script.
Fixed #2 by adding hidden inputs that post a '0' default value in case
the checkbox is unchecked. Without this, any 'on' value was not
overwritten when unchecked because unchecked checkboxes are not put into
a POST.