Rename and Restructure Project
Special thanks to Ray of OpenGarage for providing the base for this
application. MyGarage has been almost completely rewritten and differs
enough from the original that it would probably never be merged, and
can't accept merge requests from the origin anyway. Let's rename and use
it as a good time to add travis yml and restructure the files in
general.